+++ index.html
@@ -9,12 +9,15 @@
 
 	<link rel="stylesheet" href="main.css">
 	<link rel="stylesheet" href="css/fontello.css">
-
+  <!-- <script type="module">
+    import onePage from "./modules/onePage.js";
+    import openMenu from "./modules/openMenu.js";
+  </script> -->
 </head>
   <body>
 	<header class="bg-color">
 		<nav class="menu">
-				<button class="icon-menu" onclick="openMenu()" type="button"></button>
+				<button class="icon-menu" onclick="openMenu" type="button"></button>
 				<div id="Topnav" class="Topnav">
             <a class="button ative" id="defaultOpen" onclick="onePage('strona-glowna', this)" href="#strona-glowna">Strona główna</a>
             <a class="button" id="about" onclick="onePage('o-mnie', this)" href="#o-mnie">O mnie</a>
@@ -110,10 +113,11 @@
 	<footer class="bg-color">
     <div>
 		<p>Copyraight &copy; 2022 All right reserved.</p>
-		<p>Paweł Chabowski</p>
+		<p id="demo"></p>
 	</div>
 	</footer>
 	<button onclick="goTop()" id="goTop" title="Go to top"><span>↑</span></button>
-	<script src="menu.js"></script>
+
+	<script type="module" src="menu.js"></script>
 </body>
 </html>
