+++ modules/openMenu.js
@@ -1,0 +1,9 @@
+function openMenu() {
+	const menu = document.getElementById("Topnav");
+	if (menu.className == "Topnav") {
+		menu.className += " show";
+	} else {
+		menu.className = "Topnav";
+	}
+}
+export default openMenu;
